fix: minor style changes
