+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/xdd/XDDResponseOK.java
@@ -11,4 +11,4 @@ public class XDDResponseOK implements Serializable {
 	private Number v;
 
 	private String license;
-};
+}
