+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/proxies/xdd/ExtractionProxy.java
@@ -2,7 +2,10 @@ package software.uncharted.terarium.hmiserver.proxies.xdd;
 
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
-import javax.ws.rs.*;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
