+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/resources/xdd/DocumentResource.java
@@ -1,17 +1,12 @@
 package software.uncharted.terarium.hmiserver.resources.xdd;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
-
 import software.uncharted.terarium.hmiserver.proxies.xdd.DocumentProxy;
+
+import javax.ws.rs.*;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 @Path("/api/xdd/documents")
 @Produces(MediaType.APPLICATION_JSON)
@@ -19,32 +14,32 @@ import software.uncharted.terarium.hmiserver.proxies.xdd.DocumentProxy;
 @Tag(name = "XDD Documents REST Endpoint")
 public class DocumentResource {
 
-		@RestClient
-		DocumentProxy proxy;
+	@RestClient
+	DocumentProxy proxy;
 
-		// NOTE: the query parameters match the proxy version and the type XDDSearchPayload
-		@GET
-		@Produces(MediaType.APPLICATION_JSON)
-		@Tag(name = "Get all xdd documents via proxy")
-		public Response getDocuments(
-			@QueryParam("doi") String doi,
-			@QueryParam("title") String title,
-			@QueryParam("term") String term
-		) {
-			// only go ahead with the query if at least one param is present
-			if (doi != null || title != null || term != null) {
-				// for consistency, if doi/title are valid, then make sure other params are null
-				if (doi != null) {
-					title = null;
-					term = null;
-				}
-				if (title != null) {
-					doi = null;
-					term = null;
-				}
-				return proxy.getDocuments(doi, title, term);
+	// NOTE: the query parameters match the proxy version and the type XDDSearchPayload
+	@GET
+	@Produces(MediaType.APPLICATION_JSON)
+	@Tag(name = "Get all xdd documents via proxy")
+	public Response getDocuments(
+		@QueryParam("doi") String doi,
+		@QueryParam("title") String title,
+		@QueryParam("term") String term
+	) {
+		// only go ahead with the query if at least one param is present
+		if (doi != null || title != null || term != null) {
+			// for consistency, if doi/title are valid, then make sure other params are null
+			if (doi != null) {
+				title = null;
+				term = null;
 			}
-			return Response.noContent().build();
+			if (title != null) {
+				doi = null;
+				term = null;
+			}
+			return proxy.getDocuments(doi, title, term);
 		}
+		return Response.noContent().build();
+	}
 
 }
